+++ Docker/initialize.sh
@@ -2,6 +2,7 @@
 
 export HOME=/root
 export DA_ACTIVATE="${DA_PYTHON:-/usr/share/docassemble/local}/bin/activate"
+echo "Activating with ${DA_ACTIVATE}"
 source $DA_ACTIVATE
 
 export DA_CONFIG_FILE_DIST="${DA_CONFIG_FILE_DIST:-/usr/share/docassemble/config/config.yml.dist}"
@@ -505,19 +506,19 @@ if [[ $CONTAINERROLE =~ .*:(all|sql):.* ]] && [ "$PGRUNNING" = false ] && [ "$DB
     fi
 fi
 
-# echo "30"
+echo "30" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|cron):.* ]]; then
     su -c "source $DA_ACTIVATE && python -m docassemble.webapp.fix_postgresql_tables $DA_CONFIG_FILE && python -m docassemble.webapp.create_tables $DA_CONFIG_FILE" www-data
 fi
 
-# echo "31"
+echo "31" >&2
 
 if [ -f /etc/syslog-ng/syslog-ng.conf ] && [ ! -f /usr/share/docassemble/webapp/syslog-ng-orig.conf ]; then
     cp /etc/syslog-ng/syslog-ng.conf /usr/share/docassemble/webapp/syslog-ng-orig.conf
 fi
 
-# echo "32"
+# echo "32" >&2
 
 OTHERLOGSERVER=false
 
@@ -527,19 +528,19 @@ if [[ $CONTAINERROLE =~ .*:(web|celery):.* ]]; then
     fi
 fi
 
-# echo "33"
+# echo "33" >&2
 
 if [[ $CONTAINERROLE =~ .*:(log):.* ]] || [ "${LOGSERVER:-undefined}" == "null" ]; then
     OTHERLOGSERVER=false
 fi
 
-# echo "34"
+# echo "34" >&2
 
 if [ "$OTHERLOGSERVER" = false ] && [ -f ${LOGDIRECTORY:-/usr/share/docassemble/log}/docassemble.log ]; then
     chown www-data.www-data ${LOGDIRECTORY:-/usr/share/docassemble/log}/docassemble.log
 fi
 
-# echo "35"
+# echo "35" >&2
 
 if [[ $CONTAINERROLE =~ .*:(log):.* ]] || [ "$OTHERLOGSERVER" = true ]; then
     if [ -d /etc/syslog-ng ]; then
@@ -554,23 +555,23 @@ if [[ $CONTAINERROLE =~ .*:(log):.* ]] || [ "$OTHERLOGSERVER" = true ]; then
     fi
 fi
 
-# echo "36"
+echo "36" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|redis):.* ]] && [ "$REDISRUNNING" = false ]; then
     supervisorctl --serverurl http://localhost:9001 start redis
 fi
 
-# echo "37"
+echo "37" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|rabbitmq):.* ]] && [ "$RABBITMQRUNNING" = false ]; then
     supervisorctl --serverurl http://localhost:9001 start rabbitmq
 fi
 
-# echo "38"
+echo "38" >&2
 
 su -c "source $DA_ACTIVATE && python -m docassemble.webapp.update $DA_CONFIG_FILE" www-data || exit 1
 
-# echo "39"
+echo "39" >&2
 
 if su -c "source $DA_ACTIVATE && celery -A docassemble.webapp.worker status" www-data 2>&1 | grep -q `hostname`; then
     CELERYRUNNING=true;
@@ -578,19 +579,19 @@ else
     CELERYRUNNING=false;
 fi
 
-# echo "40"
+echo "40" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|celery):.* ]] && [ "$CELERYRUNNING" = false ]; then
     supervisorctl --serverurl http://localhost:9001 start celery
 fi
 
-# echo "41"
+# echo "41" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|web|log):.* ]] && [ "$APACHERUNNING" = false ]; then
     rm -f /etc/apache2/ports.conf
 fi
 
-# echo "42"
+# echo "42" >&2
 
 if [ ! -f /usr/share/docassemble/certs/apache.key ] && [ -f /usr/share/docassemble/certs/apache.key.orig ]; then
     mv /usr/share/docassemble/certs/apache.key.orig /usr/share/docassemble/certs/apache.key
@@ -609,7 +610,7 @@ if [ ! -f /usr/share/docassemble/certs/exim.crt ] && [ -f /usr/share/docassemble
 fi
 python -m docassemble.webapp.install_certs $DA_CONFIG_FILE || exit 1
 
-# echo "43"
+# echo "43" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|web):.* ]] && [ "$APACHERUNNING" = false ]; then
     echo "Listen 80" >> /etc/apache2/ports.conf
@@ -673,7 +674,7 @@ if [[ $CONTAINERROLE =~ .*:(all|web):.* ]] && [ "$APACHERUNNING" = false ]; then
     fi
 fi
 
-# echo "44"
+# echo "44" >&2
 
 if [[ $CONTAINERROLE =~ .*:(log):.* ]] && [ "$APACHERUNNING" = false ]; then
     echo "Listen 8080" >> /etc/apache2/ports.conf
@@ -681,19 +682,19 @@ if [[ $CONTAINERROLE =~ .*:(log):.* ]] && [ "$APACHERUNNING" = false ]; then
     a2ensite docassemble-log
 fi
 
-# echo "45"
+# echo "45" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|web):.* ]]; then
     supervisorctl --serverurl http://localhost:9001 start websockets
 fi
 
-# echo "46"
+# echo "46" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|web|log):.* ]] && [ "$APACHERUNNING" = false ]; then
     supervisorctl --serverurl http://localhost:9001 start apache2
 fi
 
-# echo "47"
+# echo "47" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|web):.* ]]; then
     if [ "${USEHTTPS:-false}" == "false" ]; then
@@ -707,17 +708,17 @@ if [[ $CONTAINERROLE =~ .*:(all|web):.* ]]; then
     fi
 fi
 
-# echo "48"
+# echo "48" >&2
 
 su -c "source $DA_ACTIVATE && python -m docassemble.webapp.register $DA_CONFIG_FILE" www-data
 
-# echo "49"
+# echo "49" >&2
 
 if [ "$CRONRUNNING" = false ]; then
    supervisorctl --serverurl http://localhost:9001 start cron
 fi
 
-# echo "50"
+# echo "50" >&2
 
 if [[ $CONTAINERROLE =~ .*:(all|mail):.* && ($DBTYPE = "postgresql" || $DBTYPE = "mysql") ]]; then
     if [ "${DBTYPE}" = "postgresql" ]; then
@@ -773,7 +774,7 @@ if [[ $CONTAINERROLE =~ .*:(all|mail):.* && ($DBTYPE = "postgresql" || $DBTYPE =
     supervisorctl --serverurl http://localhost:9001 start exim4
 fi
 
-# echo "51"
+# echo "51" >&2
 
 function deregister {
     su -c "source $DA_ACTIVATE && python -m docassemble.webapp.deregister $DA_CONFIG_FILE" www-data